fix bugs search function
